RE-338 Refactor TestCommand to DevCommand and split into new SubCommandRegistry + SubCommands
